Sharing cassandra session between query actor instances. Decoupled cluster construction from configuration.
